+++ catalogue_graph/src/ingestor_loader_monitor.py
@@ -6,7 +6,7 @@ from ingestor_indexer import IngestorIndexerLambdaEvent
 from models.step_events import IngestorMonitorStepEvent
 from utils.aws import pydantic_from_s3_json, pydantic_to_s3_json
 from utils.safety import validate_fractional_change
-from catalogue_graph.src.ingestor_reporter import build_final_report
+from utils.slack_report import build_final_report, LoaderReport
 
 
 class IngestorLoaderMonitorLambdaEvent(IngestorMonitorStepEvent):
@@ -21,12 +21,6 @@ class IngestorLoaderMonitorConfig(IngestorMonitorStepEvent):
     is_local: bool = False
 
 
-class LoaderReport(BaseModel):
-    pipeline_date: str
-    index_date: str
-    job_id: str
-    record_count: int
-    total_file_size: int
 
 
 def validate_events(events: list[IngestorIndexerLambdaEvent]) -> None:
@@ -75,8 +69,8 @@ def run_check(
     )
 
     s3_report_name = "report.loader.json"
-    s3_url_current_job = f"s3://{config.ingestor_s3_bucket}/{config.ingestor_s3_prefix}/{pipeline_date}/{job_id}/{s3_report_name}"
-    s3_url_latest = f"s3://{config.ingestor_s3_bucket}/{config.ingestor_s3_prefix}/{pipeline_date}/{s3_report_name}"
+    s3_url_current_job = f"s3://{config.ingestor_s3_bucket}/{config.ingestor_s3_prefix}/{pipeline_date}/{index_date}/{job_id}/{s3_report_name}"
+    s3_url_latest = f"s3://{config.ingestor_s3_bucket}/{config.ingestor_s3_prefix}/{pipeline_date}/{index_date}/{s3_report_name}"
     # Load the latest report
     latest_report = pydantic_from_s3_json(
         LoaderReport, s3_url_latest, ignore_missing=True
@@ -92,15 +86,14 @@ def run_check(
             fractional_threshold=config.percentage_threshold,
             force_pass=force_pass,
         )
+        # build and write the final pipeline report to s3 
+        build_final_report(current_report, latest_report, config)
 
     # write the current report to s3 as latest
     pydantic_to_s3_json(current_report, s3_url_latest)
 
     # write the current report to s3 as job_id
     pydantic_to_s3_json(current_report, s3_url_current_job)
-
-    # build and write the final pipeline report to s3 
-    build_final_report(current_report, latest_report, config)
 
     return current_report
