+++ catalogue_graph/tests/test_ingestor_indexer.py
@@ -3,7 +3,7 @@ from typing import Any
 
 import polars
 import pytest
-from test_mocks import MockElasticsearchClient, MockSmartOpen
+from test_mocks import MockElasticsearchClient, MockSmartOpen, MockSecretsManagerClient
 from test_utils import load_fixture
 
 from ingestor_indexer import (
@@ -17,12 +17,15 @@ from ingestor_indexer import (
 def test_ingestor_indexer_success() -> None:
     config = IngestorIndexerConfig()
     event = IngestorIndexerLambdaEvent(
+        pipeline_date="2025-01-01",
         index_date="2025-01-01",
         object_to_index=IngestorIndexerObject(
             s3_uri="s3://test-catalogue-graph/00000000-00000010.parquet"
         ),
     )
 
+    _mock_es_secrets()
+    
     # To regenerate this file after making ingestor changes, run the following command and retrieve the resulting file
     # from the `wellcomecollection-catalogue-graph` S3 bucket:
     # INGESTOR_SHARD_SIZE=10 AWS_PROFILE=platform-developer python3.13 ingestor_local.py --limit=1
@@ -36,7 +39,7 @@ def test_ingestor_indexer_success() -> None:
 
     result = handler(event, config)
     assert len(MockElasticsearchClient.inputs) == 10
-    assert result == 10  # success count
+    assert result.success_count == 10
     assert MockElasticsearchClient.inputs == expected_inputs
 
 
@@ -96,3 +99,18 @@ def test_ingestor_indexer_failure(
         MockSmartOpen.open(event.object_to_index.s3_uri, "r")
 
         handler(event, config)
+
+def _mock_es_secrets() -> None:
+    # Using a non-null pipeline_date connects to the production ES cluster, so we need to mock some secrets
+    MockSecretsManagerClient.add_mock_secret(
+        "elasticsearch/pipeline_storage_2025-01-01/private_host", "test"
+    )
+    MockSecretsManagerClient.add_mock_secret(
+        "elasticsearch/pipeline_storage_2025-01-01/port", 80
+    )
+    MockSecretsManagerClient.add_mock_secret(
+        "elasticsearch/pipeline_storage_2025-01-01/protocol", "http"
+    )
+    MockSecretsManagerClient.add_mock_secret(
+        "elasticsearch/pipeline_storage_2025-01-01/concept_ingestor/api_key", ""
+    )
